avoid overwriting primer files for same mutant

+++ scripts/T315I_ABL_design_primer.py
@@ -1,6 +1,6 @@
 ########################################
 ########################################
-
+import os.path
 from primer_design import make_single_mutant
 #from primer_design import make_mutant
 
@@ -25,11 +25,12 @@ print("Reverse Primer")
 print(reverse_primer)
 
 outfilename = "../primers/"+wt_residue+str(residue_number)+mut_residue+"_"+filename.split('/')[-1]
-with open(outfilename, 'w') as fo:
-    fo.write("Forward Primer\n")
-    fo.write(forward_primer)
-    fo.write("\nReverse Primer\n")
-    fo.write(reverse_primer)
+if not os.path.exists(outfilename):
+    with open(outfilename, 'w') as fo:
+        fo.write("Forward Primer\n")
+        fo.write(forward_primer)
+        fo.write("\nReverse Primer\n")
+        fo.write(reverse_primer)
 
 #forward_primer, reverse_primer = make_mutant(wt_sequence, wt_residue, residue_number, mut_residue, first_res=first_residue)
